Métodos listarEventos e listarAtividades movidos para EventosAdapter e AtividadeAdapter
